requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.1.13
django_injector==0.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
flake8==7.1.0
idna==3.7
injector==0.21.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
pymongo==3.12.3
pytz==2024.1
requests==2.32.3
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
